renamed role utils
